Ready for test except of ProductDiscountPolicy

+++ src/DomainLayer/Operators/ArithmeticOperators/IArithmeticOperator.cs
@@ -6,6 +6,7 @@ namespace DomainLayer.Operators.ArithmeticOperators
 {
     public interface IArithmeticOperator
     {
-        bool IsValid(int expected,int input);
+        bool IsValid(double expected,double input);
+        
     }
 }
